% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{constructFilename}
\alias{constructFilename}
\title{Construct filename from prefix and suffix}
\usage{
constructFilename(prefix, suffix = "", sep = "_", extension = ".RData")
}
\arguments{
\item{prefix}{prefix}

\item{suffix}{suffix}

\item{extension}{extension}
}
\value{
Character string of form "<prefix><sep><suffix>.<extension>"
}
\description{
Construct filename from prefix and suffix
}
